refactor(fsd-pages): replacing folders in pages via llm
Made-with: Cursor
