+++ src/pages/about/index.tsx
@@ -1,1 +1,1 @@
-export { default } from '../AboutPage/About';
+export { default } from './AboutPage/AboutPage';
