+++ src/pages/playground/PlaygroundPage/Playground.scss
@@ -1,5 +1,5 @@
-@use '../../variables.scss' as vars;
-@use '../../mixin.scss' as *;
+@use '../../../variables.scss' as vars;
+@use '../../../mixin.scss' as *;
 
 .Playground {
   width: 100%;
@@ -272,4 +272,3 @@
     max-width: 1810px;
   }
 }
-
