+++ src/pages/playground/PlaygroundPage/SettingsStage/SettingsStage.scss
@@ -1,4 +1,4 @@
-@use '../../../variables.scss' as vars;
+@use '../../../../variables.scss' as vars;
 
 .SettingsStage {
     width: 100%;
@@ -54,4 +54,3 @@
     font-size: 16px;
     display: block;
 }
-
